requirements.txt
symphony-bdk-python>=2.6.0
Jinja2~=3.0
psutil~=5.9.4